ImageLoader improved with cache functionality. Pagination implemented into ViewController

+++ DogLove/DogLove/Model/DogSearchResponse.swift
@@ -5,7 +5,7 @@
 //  Created by M. Ensar Baba on 22.11.2020.
 //
 
-import Foundation
+import UIKit
 
 
 // MARK: - DogSearchResponseElement
@@ -13,7 +13,7 @@ struct DogSearchResponseElement: Codable {
     let breeds: [Breed]?
     let id: String?
     let url: String?
-    let width, height: Int?
+    let width, height: CGFloat?
 }
 
 // MARK: - Breed
